Fixing 64 bits bug (cherry picked from commit 696b7ed391d92bb19611057e4ae9014af26177c2)
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@180494 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/j3/Compiler/JavaJIT.h
@@ -369,7 +369,7 @@ private:
                                 currentBlock);
     } else {
       llvm::Value* ttt = objectStack[currentStackIndex - 1];
-      if (((unsigned int)ttt) < 0xFFF) {
+      if (((uintptr_t)ttt) < 0xFFF) {
 
     	  printf("Why in method %s.%s???\n" , UTF8Buffer(compilingClass->name).cString(), UTF8Buffer(compilingMethod->name).cString() );
       }
